requirements.txt
flask
flask-wtf
#flask-uploads
flask-SQLAlchemy
pyshark
Flask-Reuploaded
Flask-Migrate